Refactor: Modularize database components by splitting models, engine, sessions, transactions, and mixins into separate files; update references across the codebase
