+++ src/core/database/repositories.py
@@ -9,7 +9,7 @@ from sqlalchemy.exc import IntegrityError, SQLAlchemyError
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm.strategy_options import _AbstractLoad
 
-from src.core.database.models import Base as SQLAlchemyBase
+from src.core.database.base import Base as SQLAlchemyBase
 from src.core.utils.datetime_utils import get_utc_now
 from loggers import get_logger
